renew

+++ src/com/company/ExpressionCalculator.java
@@ -1,6 +1,6 @@
 package com.company;
 
-import com.company.tokens.IntValue;
+import com.company.tokens.DoubleValue;
 import com.company.tokens.Operator;
 import com.company.tokens.Token;
 
@@ -14,23 +14,23 @@ public class ExpressionCalculator {
 
     String calculate(List<Token> tokens) {
         if (tokens.size() == 1) {
-            return Integer.toString(((IntValue) tokens.get(0)).getValue());
+            return Double.toString(((DoubleValue) tokens.get(0)).getValue());
         }
         if (tokens.size() == 3 && ((Operator) tokens.get(1)).getOperator().equals("+")) {
 
-            return Integer.toString(((IntValue) tokens.get(0)).getValue() + ((IntValue) tokens.get(2)).getValue());
+            return Double.toString(((DoubleValue) tokens.get(0)).getValue() + ((DoubleValue) tokens.get(2)).getValue());
         }
         if (tokens.size() == 3 && ((Operator) tokens.get(1)).getOperator().equals("-")) {
 
-            return Integer.toString(((IntValue) tokens.get(0)).getValue() - ((IntValue) tokens.get(2)).getValue());
+            return Double.toString(((DoubleValue) tokens.get(0)).getValue() - ((DoubleValue) tokens.get(2)).getValue());
         }
         if (tokens.size() == 3 && ((Operator) tokens.get(1)).getOperator().equals("*")) {
 
-            return Integer.toString(((IntValue) tokens.get(0)).getValue() * ((IntValue) tokens.get(2)).getValue());
+            return Double.toString(((DoubleValue) tokens.get(0)).getValue() * ((DoubleValue) tokens.get(2)).getValue());
         }
         if (tokens.size() == 3 && ((Operator) tokens.get(1)).getOperator().equals("/")) {
 
-            return Integer.toString(((IntValue) tokens.get(0)).getValue() / ((IntValue) tokens.get(2)).getValue());
+            return Double.toString(((DoubleValue) tokens.get(0)).getValue() / ((DoubleValue) tokens.get(2)).getValue());
         }
